Нода и енум вынесены в отдельные файлы
